chore: wip cdk migration schematic
