+++ libs/cdk/schematics/migration.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "../../../node_modules/@angular-devkit/schematics/collection-schema.json",
+  "schematics": {
+    "update-1-0-0": {
+      "version": "1.0.0-beta.10",
+      "description": "Migrate to sub-entry points",
+      "factory": "./src/migrations/update-1-0-0/index"
+    }
+  }
+}
