--- libs/template/schematics/src/common/consts.ts
@@ -1,14 +1,0 @@
-import {
-  NodeDependency,
-  NodeDependencyType,
-} from '@schematics/angular/utility/dependencies';
-
-export type Dependency = Omit<NodeDependency, 'version'>;
-
-export const dependencies: Dependency[] = [
-  {
-    type: NodeDependencyType.Default,
-    name: '@rx-angular/cdk',
-    overwrite: true,
-  },
-];
