--- libs/template/schematics/src/common/index.ts
@@ -1,3 +1,0 @@
-export * from './consts';
-export * from './utils/get-latest-node-version';
-export * from './utils/read-json-in-tree';
