+++ tools/generators/utils/dependency.ts
@@ -1,0 +1,4 @@
+import {
+  NodeDependency,
+} from '@schematics/angular/utility/dependencies';
+export type Dependency = Omit<NodeDependency, 'version'>;
