+++ tools/generators/utils/format-files.ts
