+++ tools/generators/utils/read-json-in-tree.ts
