+++ tools/generators/utils/replace-node-value.ts
