+++ tsconfig.base.json
@@ -21,6 +21,9 @@
       "@rx-angular/cdk/internals/scheduler": [
         "libs/cdk/internals/scheduler/src/index.ts"
       ],
+      "@rx-angular/tools/generators/*": [
+        "tools/generators/utils/*"
+      ],
       "@rx-angular/cdk/notifications": ["libs/cdk/notifications/src/index.ts"],
       "@rx-angular/cdk/render-strategies": [
         "libs/cdk/render-strategies/src/index.ts"
